refactoring duplicated code in gy/enforcement apps
